+++ src/framework/util.h
@@ -33,7 +33,7 @@ using Ref = std::reference_wrapper<T>;
 namespace Random {
 int getInt(void);
 int getInt(int max);
-float getFloat(float max);
+double getDouble(double max);
 }  // namespace Random
 
 #endif  // ALIAS_H
